refactor: split conflictResolution.ts into modular structure
Split src/terminal/conflictResolution.ts (886 lines) into focused modules:
- lwwRegister.ts (80 lines) - Last-Write-Wins register implementation
- lwwMap.ts (102 lines) - Last-Write-Wins map implementation
- textCrdt.ts (503 lines) - Text CRDT implementation
- utils.ts (42 lines) - Conflict resolution utilities
- index.ts (80 lines) - Barrel exports

Removed ConflictResolution namespace object from exports.
All functionality preserved through individual function exports.
