+++ src/terminal/conflictResolution/textCrdt.ts
@@ -1,291 +1,9 @@
 /**
- * Conflict Resolution for Collaborative Terminal Sessions
- *
- * Provides primitives for resolving concurrent edits to shared state:
- * - **LWW (Last-Write-Wins)**: For discrete state (toggles, selections, scroll positions)
- * - **Text CRDT**: For concurrent text editing (character-by-character insert/delete)
- *
- * The text CRDT uses a simplified RGA (Replicated Growable Array) approach where
- * each character has a unique ID (siteId + sequence counter). This gives
- * "good enough" collaboration semantics without full OT complexity.
- *
- * @module terminal/conflictResolution
- *
- * @example
- * ```typescript
- * import {
- *   createLWWRegister,
- *   setLWWValue,
- *   getLWWValue,
- *   createTextCRDT,
- *   insertText,
- *   deleteText,
- *   getTextValue,
- * } from 'blecsd';
- *
- * // LWW for scroll position
- * const scrollReg = createLWWRegister(0);
- * setLWWValue(scrollReg, 42, 'session-1', Date.now());
- *
- * // Text CRDT for shared text input
- * const doc = createTextCRDT('site-1');
- * insertText(doc, 0, 'Hello');
- * ```
+ * Text CRDT implementation for collaborative editing.
+ * @module terminal/conflictResolution/textCrdt
  */
 
 import { z } from 'zod';
-
-// =============================================================================
-// LWW REGISTER (Last-Write-Wins)
-// =============================================================================
-
-/**
- * Last-Write-Wins register for discrete shared state.
- * The value with the highest timestamp wins. Ties broken by site ID.
- */
-export interface LWWRegister<T> {
-	/** Current value */
-	value: T;
-	/** Timestamp of last write */
-	timestamp: number;
-	/** Site ID of last writer */
-	siteId: string;
-}
-
-/**
- * Creates a new LWW register with an initial value.
- *
- * @param initialValue - The starting value
- * @param siteId - The initial site ID (default: 'local')
- * @returns New LWW register
- *
- * @example
- * ```typescript
- * const scrollPos = createLWWRegister(0);
- * const selectedIndex = createLWWRegister(-1);
- * const isExpanded = createLWWRegister(false);
- * ```
- */
-export function createLWWRegister<T>(initialValue: T, siteId = 'local'): LWWRegister<T> {
-	return {
-		value: initialValue,
-		timestamp: 0,
-		siteId,
-	};
-}
-
-/**
- * Attempts to set the value of an LWW register.
- * The write succeeds if the timestamp is newer than the current one,
- * or if timestamps are equal and siteId is lexicographically greater (tie-breaking).
- *
- * @param register - The LWW register
- * @param value - New value
- * @param siteId - The writer's site ID
- * @param timestamp - The write timestamp
- * @returns true if the write was accepted
- *
- * @example
- * ```typescript
- * const reg = createLWWRegister(0);
- * setLWWValue(reg, 42, 'session-1', Date.now());
- * ```
- */
-export function setLWWValue<T>(
-	register: LWWRegister<T>,
-	value: T,
-	siteId: string,
-	timestamp: number,
-): boolean {
-	if (timestamp > register.timestamp) {
-		register.value = value;
-		register.timestamp = timestamp;
-		register.siteId = siteId;
-		return true;
-	}
-
-	if (timestamp === register.timestamp && siteId > register.siteId) {
-		register.value = value;
-		register.timestamp = timestamp;
-		register.siteId = siteId;
-		return true;
-	}
-
-	return false;
-}
-
-/**
- * Gets the current value of an LWW register.
- *
- * @param register - The LWW register
- * @returns Current value
- */
-export function getLWWValue<T>(register: LWWRegister<T>): T {
-	return register.value;
-}
-
-/**
- * Gets the metadata of an LWW register (timestamp and last writer).
- *
- * @param register - The LWW register
- * @returns Metadata object
- */
-export function getLWWMetadata<T>(register: LWWRegister<T>): { timestamp: number; siteId: string } {
-	return {
-		timestamp: register.timestamp,
-		siteId: register.siteId,
-	};
-}
-
-/**
- * Merges two LWW registers, returning the winning value.
- *
- * @param a - First register
- * @param b - Second register
- * @returns The register with the winning value
- */
-export function mergeLWWRegisters<T>(a: LWWRegister<T>, b: LWWRegister<T>): LWWRegister<T> {
-	if (b.timestamp > a.timestamp) {
-		return { value: b.value, timestamp: b.timestamp, siteId: b.siteId };
-	}
-	if (b.timestamp === a.timestamp && b.siteId > a.siteId) {
-		return { value: b.value, timestamp: b.timestamp, siteId: b.siteId };
-	}
-	return { value: a.value, timestamp: a.timestamp, siteId: a.siteId };
-}
-
-// =============================================================================
-// LWW MAP (for key-value state like widget properties)
-// =============================================================================
-
-/**
- * LWW Map: a map where each key has LWW semantics.
- * Useful for per-entity state like scroll positions, toggle states, etc.
- */
-export interface LWWMap<T> {
-	/** Internal map of key to LWW register */
-	readonly entries: Map<string, LWWRegister<T>>;
-}
-
-/**
- * Creates a new empty LWW map.
- *
- * @returns Empty LWW map
- *
- * @example
- * ```typescript
- * const widgetState = createLWWMap<number>();
- * setLWWMapValue(widgetState, 'scrollY', 100, 'session-1', Date.now());
- * ```
- */
-export function createLWWMap<T>(): LWWMap<T> {
-	return { entries: new Map() };
-}
-
-/**
- * Sets a value in an LWW map.
- *
- * @param map - The LWW map
- * @param key - The key
- * @param value - The value
- * @param siteId - Writer's site ID
- * @param timestamp - Write timestamp
- * @returns true if the write was accepted
- */
-export function setLWWMapValue<T>(
-	map: LWWMap<T>,
-	key: string,
-	value: T,
-	siteId: string,
-	timestamp: number,
-): boolean {
-	const existing = map.entries.get(key);
-	if (!existing) {
-		map.entries.set(key, { value, timestamp, siteId });
-		return true;
-	}
-	return setLWWValue(existing, value, siteId, timestamp);
-}
-
-/**
- * Gets a value from an LWW map.
- *
- * @param map - The LWW map
- * @param key - The key
- * @returns Value or undefined
- */
-export function getLWWMapValue<T>(map: LWWMap<T>, key: string): T | undefined {
-	const register = map.entries.get(key);
-	if (!register) {
-		return undefined;
-	}
-	return register.value;
-}
-
-/**
- * Checks if a key exists in an LWW map.
- *
- * @param map - The LWW map
- * @param key - The key
- * @returns true if exists
- */
-export function hasLWWMapKey<T>(map: LWWMap<T>, key: string): boolean {
-	return map.entries.has(key);
-}
-
-/**
- * Gets all keys in an LWW map.
- *
- * @param map - The LWW map
- * @returns Array of keys
- */
-export function getLWWMapKeys<T>(map: LWWMap<T>): readonly string[] {
-	return [...map.entries.keys()];
-}
-
-/**
- * Gets the size of an LWW map.
- *
- * @param map - The LWW map
- * @returns Number of entries
- */
-export function getLWWMapSize<T>(map: LWWMap<T>): number {
-	return map.entries.size;
-}
-
-/**
- * Merges two LWW maps. For each key present in either map,
- * the register with the winning timestamp is kept.
- *
- * @param a - First map
- * @param b - Second map
- * @returns Merged map
- */
-export function mergeLWWMaps<T>(a: LWWMap<T>, b: LWWMap<T>): LWWMap<T> {
-	const merged = createLWWMap<T>();
-
-	// Copy all from a
-	for (const [key, reg] of a.entries) {
-		merged.entries.set(key, { value: reg.value, timestamp: reg.timestamp, siteId: reg.siteId });
-	}
-
-	// Merge from b
-	for (const [key, bReg] of b.entries) {
-		const existing = merged.entries.get(key);
-		if (!existing) {
-			merged.entries.set(key, {
-				value: bReg.value,
-				timestamp: bReg.timestamp,
-				siteId: bReg.siteId,
-			});
-		} else {
-			const winner = mergeLWWRegisters(existing, bReg);
-			merged.entries.set(key, winner);
-		}
-	}
-
-	return merged;
-}
 
 // =============================================================================
 // TEXT CRDT (Simplified RGA)
@@ -789,98 +507,3 @@ export function compactDocument(doc: TextCRDT): number {
 
 	return removed;
 }
-
-// =============================================================================
-// CONFLICT RESOLUTION HELPERS
-// =============================================================================
-
-/**
- * Result of resolving a conflict between two values.
- */
-export interface ConflictResult<T> {
-	/** The winning value */
-	readonly value: T;
-	/** Which side won: 'local', 'remote', or 'merged' */
-	readonly winner: 'local' | 'remote' | 'merged';
-	/** The winning site ID */
-	readonly siteId: string;
-}
-
-/**
- * Resolves a conflict between a local and remote LWW value.
- *
- * @param localValue - Local value
- * @param localTimestamp - Local timestamp
- * @param localSiteId - Local site ID
- * @param remoteValue - Remote value
- * @param remoteTimestamp - Remote timestamp
- * @param remoteSiteId - Remote site ID
- * @returns Conflict result
- */
-export function resolveLWWConflict<T>(
-	localValue: T,
-	localTimestamp: number,
-	localSiteId: string,
-	remoteValue: T,
-	remoteTimestamp: number,
-	remoteSiteId: string,
-): ConflictResult<T> {
-	if (remoteTimestamp > localTimestamp) {
-		return { value: remoteValue, winner: 'remote', siteId: remoteSiteId };
-	}
-	if (remoteTimestamp < localTimestamp) {
-		return { value: localValue, winner: 'local', siteId: localSiteId };
-	}
-
-	// Tie: higher site ID wins
-	if (remoteSiteId > localSiteId) {
-		return { value: remoteValue, winner: 'remote', siteId: remoteSiteId };
-	}
-	return { value: localValue, winner: 'local', siteId: localSiteId };
-}
-
-/**
- * ConflictResolution namespace for convenient access.
- *
- * @example
- * ```typescript
- * import { ConflictResolution } from 'blecsd';
- *
- * const reg = ConflictResolution.createRegister(0);
- * ConflictResolution.setRegister(reg, 42, 'site-1', Date.now());
- * ```
- */
-export const ConflictResolution = {
-	// LWW Register
-	createRegister: createLWWRegister,
-	setRegister: setLWWValue,
-	getRegister: getLWWValue,
-	getRegisterMetadata: getLWWMetadata,
-	mergeRegisters: mergeLWWRegisters,
-
-	// LWW Map
-	createMap: createLWWMap,
-	setMapValue: setLWWMapValue,
-	getMapValue: getLWWMapValue,
-	hasMapKey: hasLWWMapKey,
-	getMapKeys: getLWWMapKeys,
-	getMapSize: getLWWMapSize,
-	mergeMaps: mergeLWWMaps,
-
-	// Text CRDT
-	createText: createTextCRDT,
-	getText: getTextValue,
-	getTextLength,
-	insertChar,
-	insertText,
-	deleteChar,
-	deleteText,
-	applyRemoteOp,
-	getDocumentOps,
-	getTotalNodeCount,
-	getTombstoneCount,
-	compactDocument,
-
-	// Helpers
-	resolveLWWConflict,
-} as const;
